Support default database configuration
This change allows configuring default per-profile database in
[CONFIG_DIR]/rds/default-databases.txt, where [CONFIG_DIR] is what
os.UserConfigDir [1] returns.

The file is read as `key = value` pairs, empty lines and lines starting
with # are ignored.

Key is the secrets manager entry passed in as the first argument to rds
binary, and value is the database name to automatically select.

[1]: https://pkg.go.dev/os#UserConfigDir